+++ include/character.h
@@ -18,8 +18,8 @@ class character{
     int sx, sy, dx, dy;
     int xSprites, ySprites;
     bool onGrid;
-    double typeSprite;
-    double x,y;
+    float typeSprite;
+    float x,y;
     
     sf::Texture texture;
     sf::Sprite sprite;
@@ -27,28 +27,28 @@ class character{
     public: 
     
     character();
-    character(string name, string file, double x, double y);
+    character(string name, string file, float x, float y);
     character(string file, int xi, int yi, int dir);
     character(const character& c);
     ~character(){};
     
     string getName() const;
     string getFile() const;
-    double getX() const;
-    double getY() const;
+    float getX() const;
+    float getY() const;
     int getDir() const;
-    double getTypeSprite() const;
+    float getTypeSprite() const;
     sf::Sprite getSprite() const;
     bool isOnGrid();
     sf::Vector2u getSizeSprite() const;
     
     void setName(string n);
     void setFile(string f);
-    void setX(double xNew);
-    void setY(double yNew);
+    void setX(float xNew);
+    void setY(float yNew);
     void setDir(int d);
     
-    virtual void update(double elapsedTime);
+    virtual void update(float elapsedTime);
     
     void print() const;
     character& operator=(const character& c);
@@ -59,17 +59,17 @@ class hero: public character{
     private:
     
     int wantedMove;
-    double timeKeyPressed;
-    double thresholdMove;
+    float timeKeyPressed;
+    float thresholdMove;
     bool action;
     int keyPressed;
     bool moveEnabled;
-    double speed;
-    double walk;
+    float speed;
+    float walk;
     
     public:
     
-    hero(string name, string file, double x, double y);
+    hero(string name, string file, float x, float y);
     ~hero(){};
     
     int getWantedMove() const;
@@ -83,10 +83,10 @@ class hero: public character{
     void enableMove();
     void disableMove();
     
-    void move(double elapsedTime);
-    void moveCarefully(double elapsedTime);
-    void update(double elpasedTime);
-    void updateSprite(double elapsedTime);
+    void move(float elapsedTime);
+    void moveCarefully(float elapsedTime);
+    void update(float elpasedTime);
+    void updateSprite(float elapsedTime);
     
 };
